requirements.txt
typer>=0.9.0
rich>=13.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
openai>=1.0.0
httpx>=0.24.0
filelock>=3.13.0
mcp>=1.0.0
crawl4ai>=0.4.0
fastmcp>=2.12.0
feedparser>=6.0.0
litellm>=1.57.0
tenacity==8.5.0
requests>=2.32.5